requirements.txt
PyNaCl
discord
yt-dlp
librespot
pyyaml